+++ Sources/KeyboardKit/Fakes/FakeKeyboardContext.swift
@@ -15,11 +15,6 @@ public class FakeKeyboardContext: KeyboardContext {
     
     public init() {}
     
-    @available(*, deprecated, message: "This property will be removed in KK 4.0. Usage is strongly discouraged.")
-    public var emojiCategory: EmojiCategory = .smileys
-
-    public var actionHandler: KeyboardActionHandler = FakeKeyboardActionHandler()
-    
     public var device: UIDevice = .current
     public var deviceOrientation: UIInterfaceOrientation = .portrait
     public var hasDictationKey: Bool = true