+++ Tests/KeyboardKitTests/Keyboard/ObservableKeyboardContextTests.swift
@@ -15,26 +15,19 @@ class ObservableKeyboardContextTests: QuickSpec {
     
     override func spec() {
         
-        var actionHandler: KeyboardActionHandler!
         var controller: KeyboardInputViewController!
         
         beforeEach {
-            actionHandler = MockKeyboardActionHandler()
             controller = KeyboardInputViewController()
         }
         
         describe("context") {
             
             it("can be created with params") {
-                let context = ObservableKeyboardContext(
-                    controller: controller,
-                    actionHandler: actionHandler,
-                    keyboardType: .images
-                )
+                let context = ObservableKeyboardContext(controller: controller, keyboardType: .images)
                                 
                 expect(context.device).to(be(UIDevice.current))
                 
-                expect(context.actionHandler).to(be(actionHandler))
                 expect(context.keyboardType).to(equal(.images))
                 expect(context.locale).to(equal(.current))