+++ Tests/KeyboardKitTests/System/SystemKeyboardButtonTests.swift
@@ -15,20 +15,26 @@ class SystemKeyboardButtonTests: QuickSpec {
 
     override func spec() {
         
+        var actionHandler: KeyboardActionHandler!
+        
+        beforeEach {
+            actionHandler = MockKeyboardActionHandler()
+        }
+        
         describe("system keyboard button") {
             
             it("can be created with just an action") {
-                let button = SystemKeyboardButton(action: .control)
+                let button = SystemKeyboardButton(action: .control, actionHandler: actionHandler)
                 expect(button).toNot(beNil())
             }
             
             it("can be created with a custom text") {
-                let button = SystemKeyboardButton(action: .control, text: "")
+                let button = SystemKeyboardButton(action: .control, actionHandler: actionHandler, text: "")
                 expect(button).toNot(beNil())
             }
             
             it("can be created with a custom image") {
-                let button = SystemKeyboardButton(action: .control, image: .command)
+                let button = SystemKeyboardButton(action: .control, actionHandler: actionHandler, image: .command)
                 expect(button).toNot(beNil())
             }
         }